Rename attributes file
