+++ extensions/shared/components/button/get-button-attributes.js
