+++ extensions/shared/components/button/index.js
@@ -3,6 +3,6 @@
  */
 import ButtonEdit from './edit';
 import ButtonSave from './save';
-import getButtonAttributes from './attributes';
+import getButtonAttributes from './get-button-attributes';
 
 export { ButtonEdit, ButtonSave, getButtonAttributes };
